+++ practice/src/main/java/com/spbau/bibaev/homework/vcs/command/impl/InitCommand.java
@@ -1,10 +1,9 @@
 package com.spbau.bibaev.homework.vcs.command.impl;
 
 import com.spbau.bibaev.homework.vcs.command.CommandResult;
-import com.spbau.bibaev.homework.vcs.repository.api.v2.Repository;
-import com.spbau.bibaev.homework.vcs.repository.impl.RepositoryFacade;
+import com.spbau.bibaev.homework.vcs.repository.api.Repository;
 import com.spbau.bibaev.homework.vcs.command.CommandBase;
-import com.spbau.bibaev.homework.vcs.repository.impl.v2.RepositoryImpl;
+import com.spbau.bibaev.homework.vcs.repository.impl.RepositoryImpl;
 import com.spbau.bibaev.homework.vcs.util.ConsoleColoredPrinter;
 import org.jetbrains.annotations.NotNull;
 
@@ -20,7 +19,7 @@ public class InitCommand extends CommandBase {
   @NotNull
   @Override
   protected CommandResult performImpl(@NotNull List<String> args) throws IOException {
-    if(RepositoryFacade.getInstance().openRepository(ourDirectory) != null) {
+    if(RepositoryImpl.openRepository(ourDirectory) != null) {
       ConsoleColoredPrinter.println("Repository in " + ourDirectory.toString() + " already exists");
     }
