+++ practice/src/main/java/com/spbau/bibaev/homework/vcs/command/impl/StatusCommand.java
@@ -3,9 +3,9 @@ package com.spbau.bibaev.homework.vcs.command.impl;
 import com.spbau.bibaev.homework.vcs.command.CommandResult;
 import com.spbau.bibaev.homework.vcs.command.RepositoryCommand;
 import com.spbau.bibaev.homework.vcs.repository.api.Diff;
-import com.spbau.bibaev.homework.vcs.repository.api.v2.Branch;
-import com.spbau.bibaev.homework.vcs.repository.api.v2.Commit;
-import com.spbau.bibaev.homework.vcs.repository.api.v2.Repository;
+import com.spbau.bibaev.homework.vcs.repository.api.Branch;
+import com.spbau.bibaev.homework.vcs.repository.api.Commit;
+import com.spbau.bibaev.homework.vcs.repository.api.Repository;
 import com.spbau.bibaev.homework.vcs.util.ConsoleColoredPrinter;
 import com.spbau.bibaev.homework.vcs.util.FilesUtil;
 import org.jetbrains.annotations.NotNull;
