+++ practice/src/main/java/com/spbau/bibaev/homework/vcs/repository/api/Commit.java
@@ -1,4 +1,4 @@
-package com.spbau.bibaev.homework.vcs.repository.api.v2;
+package com.spbau.bibaev.homework.vcs.repository.api;
 
 import org.jetbrains.annotations.Nullable;
