+++ practice/src/main/java/com/spbau/bibaev/homework/vcs/repository/api/Repository.java
@@ -4,36 +4,40 @@ import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
 import java.io.IOException;
+import java.nio.file.Path;
 import java.util.List;
 
-@Deprecated
 public interface Repository {
-  @NotNull
-  Project getProject();
+  WorkingDirectory getWorkingDirectory();
 
-  @NotNull
   List<Branch> getBranches();
 
-  @NotNull
   String getUserName();
 
   void setUserName(@NotNull String name) throws IOException;
 
-  @NotNull
   Branch getCurrentBranch();
 
-  @NotNull
-  Branch createNewBranch(@NotNull String name) throws IOException;
+  Branch createNewBranch(@NotNull String name, @NotNull Commit commit) throws IOException;
 
   @Nullable
   Branch getBranchByName(@NotNull String name);
 
-  @NotNull
-  Revision commitChanges(@NotNull String message) throws IOException;
+  Commit getCommit(@NotNull String commitId);
 
-  @NotNull
-  Revision checkout(@NotNull Branch branch) throws IOException;
+  RepositoryIndex getIndex();
 
-  @NotNull
-  Revision checkout(@NotNull Revision revision) throws IOException;
+  boolean addFileToIndex(@NotNull Path pathToFile);
+
+  boolean removeFileFromIndex(@NotNull Path pathToFile);
+
+  Commit commitChanges(@NotNull String message) throws IOException;
+
+  Commit merge(@NotNull Commit commit, @Nullable String message);
+
+  Commit checkout(@NotNull Branch branch) throws IOException;
+
+  Commit checkout(@NotNull Commit revision) throws IOException;
+
+  void save() throws IOException;
 }
