+++ practice/src/main/java/com/spbau/bibaev/homework/vcs/repository/api/RepositoryState.java
@@ -1,4 +1,4 @@
-package com.spbau.bibaev.homework.vcs.repository.api.v2;
+package com.spbau.bibaev.homework.vcs.repository.api;
 
 import java.util.List;
