--- practice/src/main/java/com/spbau/bibaev/homework/vcs/repository/api/v2/Branch.java
@@ -1,11 +1,0 @@
-package com.spbau.bibaev.homework.vcs.repository.api.v2;
-
-import org.jetbrains.annotations.NotNull;
-
-public interface Branch {
-  Commit getCommit();
-
-  void moveTo(@NotNull Commit commit);
-
-  String getName();
-}
