+++ practice/src/main/java/com/spbau/bibaev/homework/vcs/repository/impl/CommitImpl.java
@@ -1,9 +1,9 @@
-package com.spbau.bibaev.homework.vcs.repository.impl.v2;
+package com.spbau.bibaev.homework.vcs.repository.impl;
 
-import com.spbau.bibaev.homework.vcs.repository.api.v2.Commit;
-import com.spbau.bibaev.homework.vcs.repository.api.v2.CommitMeta;
-import com.spbau.bibaev.homework.vcs.repository.api.v2.FilePersistentState;
-import com.spbau.bibaev.homework.vcs.repository.api.v2.RepositoryState;
+import com.spbau.bibaev.homework.vcs.repository.api.Commit;
+import com.spbau.bibaev.homework.vcs.repository.api.CommitMeta;
+import com.spbau.bibaev.homework.vcs.repository.api.FilePersistentState;
+import com.spbau.bibaev.homework.vcs.repository.api.RepositoryState;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
