+++ practice/src/test/java/com/spbau/bibaev/homework/vcs/InitTest.java
@@ -1,7 +1,7 @@
 package com.spbau.bibaev.homework.vcs;
 
-import com.spbau.bibaev.homework.vcs.repository.api.v2.Repository;
-import com.spbau.bibaev.homework.vcs.repository.impl.v2.RepositoryImpl;
+import com.spbau.bibaev.homework.vcs.repository.api.Repository;
+import com.spbau.bibaev.homework.vcs.repository.impl.RepositoryImpl;
 import org.junit.Test;
 
 import java.io.IOException;
