+++ practice/src/test/java/com/spbau/bibaev/homework/vcs/command/impl/AddCommandTest.java
@@ -3,7 +3,7 @@ package com.spbau.bibaev.homework.vcs.command.impl;
 import com.spbau.bibaev.homework.vcs.RepositoryTestCase;
 import com.spbau.bibaev.homework.vcs.command.Command;
 import com.spbau.bibaev.homework.vcs.command.CommandResult;
-import com.spbau.bibaev.homework.vcs.repository.api.v2.Repository;
+import com.spbau.bibaev.homework.vcs.repository.api.Repository;
 import org.jetbrains.annotations.Nullable;
 import org.junit.Test;
